Extract annotation to special package
